move gamestate to its own file and connect it to App
